loadImage move to Loader
